fix: address OpenAPI spec bugs and remove unsafe type casts
- Normalize GET /api/foods?barcode response to { foods, total } shape
  (was returning { food } which broke barcode scanning in DayLog)
- Fix search query param name from 'search' to 'q' in OpenAPI spec
  (food search was silently broken)
- Add oneOf for GET /api/weight to model both entries and trend responses
  (was only declaring entries shape, hiding trend response from types)
- Add requestBody for POST /api/supplements/{id}/log
- Add force query param to DELETE /api/recipes/{id} spec
- Fix FavoriteMealTimeframe schema to include all server-returned fields
  (id, startMinute, endMinute, sortOrder were missing)
- Fix OpenFoodFacts nutriScore to use enum type matching food schema
- Remove 12 'as any' casts from migrated frontend code
- Migrate recipes confirmForceDelete from apiFetch to typed client
- Regenerate OpenAPI spec, TypeScript types, and Kotlin DTOs

+++ src/lib/server/validation/responses/favorites.ts
@@ -11,7 +11,7 @@ const favoriteFoodSchema = z
 		carbs: z.number(),
 		fat: z.number(),
 		fiber: z.number(),
-		logCount: z.number(),
+		logCount: z.number().int(),
 		type: z.literal('food')
 	})
 	.meta({ id: 'FavoriteFood' });
@@ -26,7 +26,7 @@ const favoriteRecipeSchema = z
 		carbs: z.number(),
 		fat: z.number(),
 		fiber: z.number(),
-		logCount: z.number(),
+		logCount: z.number().int(),
 		totalServings: z.number(),
 		type: z.literal('recipe')
 	})
